Integrated 3D into Galaxy Engine.
Also:
Fixed some hot-reload bugs.
Fixed gamma issues in 2D rendering shaders.
Fixed some move operator/constructor issues in Shader class.
Updated Lua script functions for engine.
Fixed duplicate key inputs in 2D sandbox.
Fixed other misc issues.

+++ galaxy/src/galaxy/ecs/ComponentSet.hpp
@@ -193,7 +193,7 @@ namespace galaxy
 		template<meta::is_class Component>
 		inline const unsigned int ComponentSet<Component>::get_size() const noexcept
 		{
-			return m_components.size();
+			return static_cast<unsigned int>(m_components.size());
 		}
 
 		template<meta::is_class Component>
